Project2_CDC_ETL_CLI/requirement.txt
requests
pandas
mysql-connector-python
argparse
configparser
